+++ sample/label.cpp
@@ -28,7 +28,7 @@ public:
     L(L2);
     ret();
   }
-  const uint32_t *gen() {
+  const uint8_t *gen() {
     genAddFunc();
     ready();
     return getCode();
